+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/minimal/ui/StoreDemo.java
@@ -24,7 +24,6 @@ import is.codion.dbms.h2.H2DatabaseFactory;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.demos.manual.store.minimal.domain.Store;
-import is.codion.swing.common.ui.component.table.ConditionPanel.ConditionState;
 import is.codion.swing.common.ui.dialog.Dialogs;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 import is.codion.swing.framework.model.SwingEntityModel;
@@ -38,6 +37,7 @@ import javax.swing.UIManager;
 
 import static is.codion.framework.demos.manual.store.minimal.domain.Store.Address;
 import static is.codion.framework.demos.manual.store.minimal.domain.Store.Customer;
+import static is.codion.swing.common.ui.component.table.ConditionPanel.ConditionView.SIMPLE;
 import static is.codion.swing.common.ui.layout.Layouts.gridLayout;
 import static javax.swing.BorderFactory.createEmptyBorder;
 
@@ -113,7 +113,7 @@ public class StoreDemo {
 		customerPanel.addDetailPanel(addressPanel);
 
 		addressPanel.tablePanel()
-						.conditions().state().set(ConditionState.SIMPLE);
+						.conditions().conditionView().set(SIMPLE);
 
 		customerModel.tableModel().refresh();
 		customerPanel.setBorder(createEmptyBorder(5, 5, 0, 5));