+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -40,8 +40,8 @@
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#WINDOW_TYPE}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#QUERY_HIDDEN_COLUMNS}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#ORDER_QUERY_BY_SORT_ORDER}
- * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#CONDITION_STATE}
- * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#FILTER_STATE}
+ * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#CONDITION_VIEW}
+ * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#FILTER_VIEW}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#SUMMARY_PANEL_VISIBLE}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_POPUP_MENU}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#INCLUDE_ENTITY_MENU}