+++ changelog.md
@@ -27,6 +27,7 @@ Codion Change Log
 - ColumnConditionPanel renamed ConditionPanel.
 - ColumnConditionsPanel.panels() and panel() renamed get().
 - ColumnConditionsPanel renamed TableConditionPanel.
+- ConditionPanel.ConditionState renamed ConditionView.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.